move JoinLobby
